Move raylib logger redirect in raylib namespace
Resolves #93
