+++ src/raylib/raylib.hpp
@@ -1,0 +1,21 @@
+/*
+** EPITECH PROJECT, 2022
+** Bomberman
+** File description:
+** raylib
+*/
+
+#ifndef RAYLIB_HPP_
+#define RAYLIB_HPP_
+
+namespace raylib
+{
+    /// Redirect the raylib logger in log_raylib.txt.
+    ///
+    /// @param logLevel loglevel of the raylib.
+    ///
+    void initLogger(int logLevel);
+
+} // namespace raylib
+
+#endif /* !RAYLIB_HPP_ */
